Be 6 add authorization

+++ app/services/user.py
@@ -11,6 +11,7 @@ from app.schemas.user import (
 from app.core.security import hash_password, verify_password
 from app.repositories.user import UserRepository
 from app.core.exceptions import (
+  ForbiddenAction,
   UserNotFound,
   ExistsEmail,
   ExistsUsername,
@@ -54,15 +55,21 @@ class UserService:
     self,
     email: str
   ) -> User | None:
+    user = await self.repo.get_user_by_email(email)
+    if not user:
+      raise UserNotFound()
     logger.info(f"Fetched user email={email}")
-    return await self.repo.get_user_by_email(email)
+    return user
 
   async def get_user_by_provider_id(
     self,
     provider_id: str
   ) -> User | None:
+    user = await self.repo.get_user_by_provider_id(provider_id)
+    if not user:
+      raise UserNotFound()
     logger.info(f"Fetched user provider_id={provider_id}")
-    return await self.repo.get_user_by_provider_id(provider_id)
+    return user
 
   async def create_new_user(
     self,
@@ -104,9 +111,17 @@ class UserService:
 
   async def update_user_details(
     self,
-    user: User,
-    update_data: UserUpdate
+    user_id: UUID,
+    update_data: UserUpdate,
+    current_user_id: UUID
   ) -> UserDetailResponse:
+    # Переконуємося що користвуач змінює свої дані
+    if user_id != current_user_id:
+      raise ForbiddenAction("You are trying to edit data that is not yours")
+    user = await self.repo.get_user_by_id(user_id)
+    if not user:
+      raise UserNotFound()
+    logger.info(f"Fetched user id={user_id}")
     # --- Username ---
     if update_data.username is not None:
       # якщо username реально змінюється
@@ -134,7 +149,16 @@ class UserService:
 
   async def delete_user(
     self,
-    user: User
+    user_id: UUID,
+    current_user_id: UUID
   ) -> None:
+    # Переконуємося що користвуач змінює свої дані
+    if user_id != current_user_id:
+      raise ForbiddenAction("You are trying to edit data that is not yours")
+    user = await self.repo.get_user_by_id(user_id)
+    if not user:
+      raise UserNotFound()
+    logger.info(f"Fetched user id={user_id}")
     await self.repo.delete_user(user)
-    logger.info(f"User deleted id={user.id}")
+    logger.info(f"User deleted id={user_id}")
+    return
